Refactor, restructure
Restructure the code so that statuses and severities are persisted with
JSON files and stored in a ConfigData class. Create a Preferences class
to store the author name and some (future) GUI preferences. Remove
classes, Services, and ServiceImpl’s for Author, Status, and Severity.
Refactor the test database code to reflect all changes. A few
additional minor tweaks.